Lab15_Auth_JWT/requirements.txt
django==3.2
djangorestframework
django-cors-headers
djangorestframework_simplejwt